Fixed major bug causing crash due to notification center
**Mem warnings are being thrown from the menu, may be due to the
allocation of SKScenes in the collectionView

+++ dropBeats/InstrumentNode.h
@@ -19,7 +19,7 @@
 @property (nonatomic) DPNote* note;
 @property (nonatomic) NoteType instrumentNoteIndex;
 
--(id)initWithInstrumentIndex: (int) index andSize: (CGSize) size;
+-(id)initWithInstrumentIndex: (NSInteger) index andSize: (CGSize) size;
 +(void)loadActions;
 -(void)playInstrument;
